+++ include/render/video_renderer.h
@@ -42,6 +42,9 @@ public:
     virtual bool consumeSpeedChangeRequest(double& speed_delta) = 0;
     virtual bool consumeResetSpeedRequest() = 0;
     virtual bool consumeToggleSubtitleRequest() = 0;
+    virtual bool consumeSetABRepeatStartRequest() = 0;
+    virtual bool consumeSetABRepeatEndRequest() = 0;
+    virtual bool consumeClearABRepeatRequest() = 0;
     virtual bool consumeNextChapterRequest() = 0;
     virtual bool consumePreviousChapterRequest() = 0;
     virtual bool consumeNextItemRequest() = 0;