+++ src/input/hotkey_manager.cpp
@@ -19,7 +19,7 @@ namespace {
 
 using ActionPair = std::pair<PlayerAction, const char*>;
 
-constexpr std::array<ActionPair, 16> kActionKeyPairs{{
+constexpr std::array<ActionPair, 19> kActionKeyPairs{{
     {PlayerAction::PlayPause, "play_pause"},
     {PlayerAction::SeekBackward, "seek_backward"},
     {PlayerAction::SeekForward, "seek_forward"},
@@ -29,6 +29,9 @@ constexpr std::array<ActionPair, 16> kActionKeyPairs{{
     {PlayerAction::SpeedDown, "speed_down"},
     {PlayerAction::SpeedUp, "speed_up"},
     {PlayerAction::SpeedReset, "speed_reset"},
+    {PlayerAction::SetABRepeatStart, "ab_repeat_start"},
+    {PlayerAction::SetABRepeatEnd, "ab_repeat_end"},
+    {PlayerAction::ClearABRepeat, "ab_repeat_clear"},
     {PlayerAction::PreviousChapter, "previous_chapter"},
     {PlayerAction::NextChapter, "next_chapter"},
     {PlayerAction::PreviousItem, "previous_item"},
@@ -105,6 +108,9 @@ void HotkeyManager::resetToDefaults() {
     bind(PlayerAction::SpeedDown, SDLK_LEFTBRACKET);
     bind(PlayerAction::SpeedUp, SDLK_RIGHTBRACKET);
     bind(PlayerAction::SpeedReset, 'r');
+    bind(PlayerAction::SetABRepeatStart, 'a');
+    bind(PlayerAction::SetABRepeatEnd, 'b');
+    bind(PlayerAction::ClearABRepeat, 'c');
     bind(PlayerAction::PreviousChapter, SDLK_HOME);
     bind(PlayerAction::NextChapter, SDLK_END);
     bind(PlayerAction::PreviousItem, SDLK_PAGEUP);